The intended interface

IJuliaTimeMachine.@past — Macro

@past n

Return to the state after cell n was exectued.

IJuliaTimeMachine.vars — Function

vars(cell)
vars(vx::Varchive, cell)

Returns a dictionary of the variables from cell cell. If vx is omitted, returns from the default history. This is how it is usually used.

IJuliaTimeMachine.dontsave — Function

dontsave(x)

Do not save variable x in the history.

IJuliaTimeMachine.ans — Function

Recall the answer from a cell

IJuliaTimeMachine.@thread — Macro

TM.@thread begin
    code you want to run
end

Spawns a process that runs your code in its own thread, and eventually saves the result when it finishes in Out[cellnum]. It copies variables, so that changes to those variables do not interact with other cells.

IJuliaTimeMachine.notify_jupyter! — Function

notify_jupyter!(bool)

If true, Print notifications about finishing jobs to the current jupyter output cell.

IJuliaTimeMachine.notify_terminal! — Function

notify_terminal!(bool)

If true, Print notifications about finishing jobs to the terminal from which jupyter was started.

IJuliaTimeMachine.unhook — Function

unhook()

Remove the pre and postexecute hooks created by Time Machine. Made for when the Time Machine is causing strange errors. These typically cause strange errors to happen whenever a cell is executed.

All Docstrings

IJuliaTimeMachine.ans — Method

Recall the answer from a cell

IJuliaTimeMachine.can_copy — Method

can_copy(x)

Is supposed to return true if deepcopy works on x. The code to test this is based on deepcopy, but it could get some strange case wrong.

IJuliaTimeMachine.clear_past — Method

clear_past(indices)

Empty all storage of the past. Use to free up memory. If indices is omitted, it clears all history.

IJuliaTimeMachine.dontsave — Method

dontsave(x)

Do not save variable x in the history.

IJuliaTimeMachine.extract_symbols — Method

The goal of extract_symbols is to find all the variables that appear in a block. It returns them in a Set of Symbols. This will return many more symbols than we want, so we only save those that we can copy.

IJuliaTimeMachine.let_block — Method

Creates the let statement for all copyable variables.

IJuliaTimeMachine.main_to_dict — Method

Go over every symbol in Main. If deepcopy works on it, put it in the dict d. Note that it doesn't capture functions, which is unfortunate.

IJuliaTimeMachine.notify_jupyter! — Function

notify_jupyter!(bool)

If true, Print notifications about finishing jobs to the current jupyter output cell.

IJuliaTimeMachine.notify_terminal! — Function

notify_terminal!(bool)

If true, Print notifications about finishing jobs to the terminal from which jupyter was started.

IJuliaTimeMachine.put_state! — Method

put_state!(vx::Varchive, key, di::Dict, ansc)

Assuming that state indexed by key (=n) is packed into ijstate, put it into the Varchive.

IJuliaTimeMachine.save_state — Method

Save the current variables and ans in VX (a Varchive).

IJuliaTimeMachine.save_state_block — Method

Creates a block that when executed saves the used variables in VX.

IJuliaTimeMachine.saving! — Function

saving!(bool)

Turn saving of state on or off. Works by pushing or poping an IJulia postexecute_hook. True by default.

IJuliaTimeMachine.tm_hash — Method

tm_hash(x)

Produces a hash of any variable for which can_copy returns true. If variables x and y are different, they will probably have different hashes.

IJuliaTimeMachine.unhook — Method

unhook()

Remove the pre and postexecute hooks created by Time Machine. Made for when the Time Machine is causing strange errors. These typically cause strange errors to happen whenever a cell is executed.

IJuliaTimeMachine.vars — Method

vars(cell)
vars(vx::Varchive, cell)

Returns a dictionary of the variables from cell cell. If vx is omitted, returns from the default history. This is how it is usually used.

IJuliaTimeMachine.@dict_to_main — Macro

@dict_to_main dict

Place all variables in dict into Main memory.

IJuliaTimeMachine.@past — Macro

@past n

Return to the state after cell n was exectued.

IJuliaTimeMachine.@thread — Macro

TM.@thread begin
    code you want to run
end

Spawns a process that runs your code in its own thread, and eventually saves the result when it finishes in Out[cellnum]. It copies variables, so that changes to those variables do not interact with other cells.